Move look assignment logic to dedicated looks lib
