+++ openpype/hosts/maya/tools/mayalookassigner/app.py
@@ -8,7 +8,7 @@ from openpype.client import get_last_version_by_subset_id
 from openpype import style
 from openpype.pipeline import legacy_io
 from openpype.tools.utils.lib import qt_app_context
-from openpype.hosts.maya.api.lib import assign_look_by_version
+from openpype.hosts.maya.api.lib_looks import assign_look_by_version
 
 from maya import cmds
 # old api for MFileIO
